refactor(Clients): Move 'get_scope' from bin_legacy to utils
Issue: #8298
