+++ lib/rucio/cli/bin_legacy/rucio.py
@@ -36,7 +36,7 @@ from tabulate import tabulate
 
 # rucio module has the same name as this executable module, so this rule fails. pylint: disable=no-name-in-module
 from rucio import version
-from rucio.cli.utils import exception_handler, get_client, setup_gfal2_logger, signal_handler
+from rucio.cli.utils import exception_handler, get_client, get_scope, setup_gfal2_logger, signal_handler
 from rucio.client.richclient import MAX_TRACEBACK_WIDTH, MIN_CONSOLE_WIDTH, CLITheme, generate_table, get_cli_config, get_pager, print_output, setup_rich_logger
 from rucio.common.client import detect_client_location
 from rucio.common.config import config_get, config_get_float
@@ -53,7 +53,7 @@ from rucio.common.exception import (
 )
 from rucio.common.extra import import_extras
 from rucio.common.test_rucio_server import TestRucioServer
-from rucio.common.utils import Color, StoreAndDeprecateWarningAction, chunks, extract_scope, parse_did_filter_from_string, parse_did_filter_from_string_fe, setup_logger, sizefmt
+from rucio.common.utils import Color, StoreAndDeprecateWarningAction, chunks, parse_did_filter_from_string, parse_did_filter_from_string_fe, setup_logger, sizefmt
 
 if TYPE_CHECKING:
     from rucio.common.types import FileToUploadDict
@@ -71,17 +71,6 @@ DEFAULT_PORT = 80
 
 tablefmt = 'psql'
 cli_config = get_cli_config()
-
-
-def get_scope(did, client):
-    try:
-        scope, name = extract_scope(did)
-        return scope, name
-    except TypeError:
-        scopes = client.list_scopes()
-        scope, name = extract_scope(did, scopes)
-        return scope, name
-    return None, did
 
 
 def __resolve_containers_to_datasets(scope, name, client):
