+++ lib/rucio/cli/bin_legacy/rucio_admin.py
@@ -33,7 +33,7 @@ from rich.tree import Tree
 from tabulate import tabulate
 
 from rucio import version
-from rucio.cli.utils import exception_handler, get_client, setup_gfal2_logger, signal_handler
+from rucio.cli.utils import exception_handler, get_client, get_scope, setup_gfal2_logger, signal_handler
 from rucio.client.richclient import MAX_TRACEBACK_WIDTH, MIN_CONSOLE_WIDTH, CLITheme, generate_table, get_cli_config, get_pager, print_output, setup_rich_logger
 from rucio.common.constants import RseAttr
 from rucio.common.exception import (
@@ -44,7 +44,7 @@ from rucio.common.exception import (
     RucioException,
 )
 from rucio.common.extra import import_extras
-from rucio.common.utils import StoreAndDeprecateWarningAction, chunks, clean_pfns, construct_non_deterministic_pfn, extract_scope, get_bytes_value_from_string, parse_response, render_json, setup_logger, sizefmt
+from rucio.common.utils import StoreAndDeprecateWarningAction, chunks, clean_pfns, construct_non_deterministic_pfn, get_bytes_value_from_string, parse_response, render_json, setup_logger, sizefmt
 from rucio.rse import rsemanager as rsemgr
 
 EXTRA_MODULES = import_extras(['argcomplete'])
@@ -64,16 +64,6 @@ DEFAULT_PORT = 443
 
 tablefmt = 'psql'
 cli_config = get_cli_config()
-
-
-def get_scope(did, client):
-    try:
-        scope, name = extract_scope(did)
-        return scope, name
-    except TypeError:
-        scopes = client.list_scopes()
-        scope, name = extract_scope(did, scopes)
-        return scope, name
 
 
 @exception_handler
